Run lint fix
