+++ frontend/src/components/buttons/SubmitContactFormButton.jsx
@@ -7,7 +7,9 @@ export default function SubmitContactFormButton({ onClick }) {
 			className="bg-mv-green border-solid border-mv-green rounded-[5px] w-full h-[69px] shadow-buttonShadow"
 			onClick={onClick}
 		>
-			<div className="text-mv-white font-poppins text-lg font-medium leading-20 tracking-tight text-center">Submit</div>
+			<div className="text-mv-white font-poppins text-lg font-medium leading-20 tracking-tight text-center">
+				Submit
+			</div>
 		</button>
 	);
 }
